Passport integration & looottss of fixes

+++ src/main/kotlin/com/scrumdapp/groupservice/dto/AddUserDto.kt
@@ -1,5 +1,5 @@
 package com.scrumdapp.groupservice.dto
 
 data class AddUserDto(
-    val userId: Int
+    val userId: Long
 )
