+++ src/main/kotlin/com/scrumdapp/groupservice/entities/Group.kt
@@ -21,7 +21,7 @@ class Group {
     var name: String? = null
 
     @Column(nullable = false)
-    var group_owner: Int = 0
+    var group_owner: Long = 0
 
     @Column(nullable = true)
     var background_preference: Int? = null
